# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Teacoder educational platform",
        "title": "TeaCoder API",
        "contact": {
            "name": "TeaCoder Team",
            "url": "https://teacoder.ru",
            "email": "help@teacoder.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/auth/account/create": {
            "post": {
                "description": "Данный метод создает нового пользователя на основе предоставленных данных.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Создать нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для создания нового пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно создан",
                        "schema": {
                            "$ref": "#/definitions/routes.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные для создания пользователя",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Пользователь с таким email уже существует",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "routes.UserDTO": {
            "description": "Данные, используемые для создания нового пользователя",
            "type": "object",
            "required": [
                "display_name",
                "email",
                "password"
            ],
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Иван Иванов"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}